feat: use the raw homeserver URL instead of manually removing the scheme
On the one hand, the SDK knows what to do and will clean up the URL if needs be,
guessing what the best scheme (http or https) is based on the prefix. On the other
hand, trimming the scheme as done here doesn't allow for manual testing on insecure
servers, which is handy in e2e testing situations.