TimelineItemPresenterFactories
DI infrastructure to allow injection of presenters into the timeline.
Add an @AssisteFactory of type `TimelineItemPresenterFactory` to a [Presenter] class and bind this factory ino the TimelineItemPresenterFactory map multi binding using:

```
@Binds
@IntoMap
@TimelineItemEventContentKey(MyTimelineItemContent::class)
```

A map multibinding of such map will be available in the `LocalTimelineItemPresenterFactories` composition local for further use down the UI tree.