Split `VoiceMessageType` from `AudioMessageType`
Currently, for compatibility reasons, we implement MSC3245v1 which puts the voice data inside an audio message type.
Though at times it seems impractical to deal with a single message type which effectively represents 2 different kinds of messages.
This create a new message type called `VoiceMessageType` which is used whenever we receive an event with `"msgtype": "m.audio"` which also has the `"org.matrix.msc3245.voice": {}` field.
This makes it easier to process voice messages as different entities throughout the codebase.