Bug: rememberSaveable on timeline items behaves erratically
Reproduce steps:
Launch the app go to a room with many voice messages.
Click play on some of them, it will increase the counters displayed within the voice message.
Scroll the timeline untile the voice messages go out of view.
You will see that the 2 digits (the saved and not saved reset at different times but eventually both of them reset).