Simple Html page to view all the screenshots.
Do not delete the whole screenshots folder, but just delete per language when the screenshots are generated.

Generate the data.js file regarding the existing screenshots, to be able to add languages.
Also the language will be sorted alphabetically (English still first).
And the screenshot will be sorted alphabetically.

Update the script to generate screenshot for specific language on demand.