Fix dependencyAnalysis plugin.
For some reason the plugin needs to be manually applied to all the libs, otherwise running `./gradlew buildHealth` does nothing.

Also, I had to add it to the convention plugins instead of using the `subprojects { ... }` approach because of the `:libraries:rustsdk` and `:libraries:textcomposer:lib` modules, which aren't proper Java/Kotlin modules.