Fix crash when using Element Call on API <= 30
`commsDeviceChangedListener` is added as a non-null property but the `AudioManager.OnCommunicationDeviceChangedListener` class it implements is only available on API 31+. Although it's not used in APIs <= 30 the property definition is still there and the app will crash when trying to run the code.

Making it a `lazy` property works around the issue.