Disable BT audio devices for Element Call on Android <= 11
After trying several strategies to make BT audio devices play nice with the embedded Element Call instance on Android versions using the legacy audio APIs - Android 11 and lower - and it failing in several ways - wrong audio device being used when switching, audio just dying until the app is force closed or the app crashing with a native OS error and no stacktrace - we had to make a difficult decision of preventing using these devices at all.

This change makes it so when selecting a BT audio device or when it's selected by default by the OS, we notify the user about the device not being usable, and mute the audio to prevent it from being streamed through some other audio device. Once a new device is selected, the audio is enabled again and that device can be used normally.